+++ src/typey-boi/TypeyBoi.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import { WordInput } from './components/WordInput'
+
+ReactDOM.render(<WordInput />, document.getElementById('root'))