+++ src/typey-boi/typey-boi.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react'
-import * as ReactDOM from 'react-dom'
-
-ReactDOM.render(
-  <p>Content goes here</p>,
-  document.getElementById('root'),
-)