+++ CHANGELOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Added boilerplate for project.
  - Added electron packaging.
+ - Added ability to classify typed input as correct/incorrect.